+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - [#1280] AS3 Direct editation of traits with the same name
 - [#1743] GFX - Adding DefineExternalImage2 and DefineSubImage tags
 - [#1822], [#1803] AS3 direct editation - optional using AIR (airglobal.swc) to compile
+- [#1501] Bulk import shapes
 
 ### Fixed
 - [#1869] Replace references now replaces all references, not just PlaceObject
@@ -30,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - [#1840] AS3 Direct editation - Type mismatched for a trait
 - [#1840] Proper if..continue..break handling
 - [#1877] Recalculate dependent characters and frames on removing / editing item
+- DefineShape4 SVG import NullPointerException
 
 ### Changed
 - GFX - DefineExternalImage2 no longer handled as character
@@ -2611,6 +2613,7 @@ All notable changes to this project will be documented in this file.
 [#1743]: https://www.free-decompiler.com/flash/issues/1743
 [#1822]: https://www.free-decompiler.com/flash/issues/1822
 [#1803]: https://www.free-decompiler.com/flash/issues/1803
+[#1501]: https://www.free-decompiler.com/flash/issues/1501
 [#1869]: https://www.free-decompiler.com/flash/issues/1869
 [#1872]: https://www.free-decompiler.com/flash/issues/1872
 [#1692]: https://www.free-decompiler.com/flash/issues/1692